Driver Class
The core of the classes. This reads in the dictionary file wordsEN.txt and converts it using the Alphabetize class adding all the words and keys to a global HashMap. Then we take user input and convert it to all possible keys using Alphabits and search our map for key matches. The Map returns sets containing possible words and we finally print these words out.